# openapi.yaml
openapi: 3.0.3
info:
  title: Beacon API
  version: "1.0.0"
  description: |-
    Beacon is a crisis-damage crowdsourcing backend: anonymous mobile reporters
    submit building-damage reports (identified only by an `X-Device-Id` header),
    and authenticated analysts (JWT bearer) verify, dispatch and export them.

    ## Visibility tiers
    - **Anonymous / public** callers (no JWT) and the authenticated read-only
      `external_viewer` role see **verified reports only**, in a locked-down
      *public projection*: coordinates coarsened to 3 decimal places (~110 m),
      and `submitterId`, `what3words`, `plusCode`, `landmark`, `buildingId`,
      `buildingSource`, `infraName`, `gpsAccuracyMeters`, `description`,
      tasking/operational fields, `modular` and `anonymization` are stripped.
      Unverified reports return 404 so their existence and position are never
      leaked. Public aggregates (`/reports/area-groups`) count verified
      reports only for this tier, coherent with the verified-only public map.
    - **Analyst roles** (`field_validator`, `co_analyst`, `regional_analyst`,
      `crisis_admin`) see all statuses at full precision, limited to the
      crises in their `crisisScope` (`"*"` = org-wide).

    ## Crisis scoping
    Where `crisisId` is optional (stats, export), an omitted value resolves to
    the **newest active crisis** (the same crisis `GET /v1/crises/active`
    returns), falling back to a seeded default only when no crisis is active.

...

...

...
servers:
...
paths:
...
  /v1/reports/{id}/verification:
    patch:
      tags: [reports]
      operationId: patchVerification
      summary: Set a report's verification status (analyst decision)
      description: |-
        Mutator endpoint — any analyst role except the read-only
        `external_viewer` (403 `forbidden`), crisis-scoped: an analyst may
        only verify reports of crises in their scope (403 `out_of_scope`); a
        pending report with no crisis is mutable only with org-wide scope.

        **Photo gate:** setting `status: verified` on a report with NO photo
        is rejected with 409 `photo_required` unless `force: true` is passed.
        "Photo" means evidence actually **uploaded to the server** (a stored
        image, served by `GET /v1/reports/{id}/photo`) — photo metadata merely
        claimed by the client in the submit body does not count. The optional
        `note` and the `force` flag are recorded in the verification audit
        trail.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ReportId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerificationRequest'
      responses:
        "200":
          description: The updated report.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Report'
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/OutOfScope'
        "404":
          $ref: '#/components/responses/NotFound'
        "409":
          description: |-
            Verifying a photo-less report without `force=true`
            (`photo_required`). "Photo-less" = no evidence uploaded to the
            server; client-claimed photo metadata does not satisfy the gate.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Analyst JWT issued by POST /v1/auth/login.
  parameters:
    ReportId:
      name: id
      in: path
      required: true
      schema: { type: string }
      description: The report id (client-generated at submit time).
...
  responses:
    BadRequest:
      description: Malformed request.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Analyst login required (missing/invalid bearer token).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    OutOfScope:
      description: The caller's role/scope cannot access or modify this resource (`out_of_scope` / `forbidden`).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found (also returned for resources the caller's tier may not see).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: Stable machine-readable code (e.g. validation, bad_json, unauthorized, out_of_scope, not_found, rate_limited).
        message:
          type: string
...
    PhotoRef:
      type: object
      properties:
        localPath: { type: string }
        remoteUrl:
          type: string
          nullable: true
        sizeBytes:
          type: integer
          format: int64
    ReportLocation:
      type: object
      properties:
        lat:
          type: number
          nullable: true
          description: Null for a location-unresolved (landmark-only) report.
        lng:
          type: number
          nullable: true
        buildingId:
          type: string
          nullable: true
        buildingSource:
          type: string
          nullable: true
          description: '"footprint" only when a real footprint polygon was tapped; null otherwise.'
        what3words:
          type: string
          nullable: true
          description: Legacy misnamed Plus Code alias (same value as plusCode, kept for compat).
        plusCode:
          type: string
          nullable: true
        landmark:
          type: string
          nullable: true
        gpsAccuracyMeters:
          type: number
          nullable: true
    ReportDescription:
      type: object
      properties:
        original: { type: string }
        originalLang: { type: string }
        translated:
          type: string
          description: Always emitted in responses; coalesces to the original text when no translation exists.
        translatedLang:
          type: string
          nullable: true
    AdminRef:
      type: object
      description: "One administrative level. `pcode` is a geoBoundaries shapeID (`GB:<shapeID>`) or an illustrative seed code — NOT an official OCHA COD P-code."
      properties:
        pcode: { type: string }
        name: { type: string }
    AdminChain:
      type: object
      description: Reverse-geocoded ADM0–ADM3 chain stamped on a report (same provenance caveat as AdminRef).
      properties:
        adm0: { $ref: '#/components/schemas/AdminRef' }
        adm1: { $ref: '#/components/schemas/AdminRef' }
        adm2: { $ref: '#/components/schemas/AdminRef' }
        adm3: { $ref: '#/components/schemas/AdminRef' }
    Anonymization:
      type: object
      description: |-
        Honest privacy flags. exifStripped is real (mobile strips EXIF on
        capture); facesBlurred/platesBlurred are NOT implemented and the
        server forces them to false even if a client claims otherwise.
      properties:
        anonymous: { type: boolean }
        exifStripped: { type: boolean }
        facesBlurred: { type: boolean }
        platesBlurred: { type: boolean }
    ModularImpacts:
      type: object
      description: |-
        Modular secondary-impact answers. The server stores this blob
        verbatim (no strict server-side enum validation); the enums below are
        the documented wire contract, and the values are flattened into the
        export columns electricity / health_services / pressing_needs.
      properties:
        electricity:
          type: string
          nullable: true
          enum: [none_observed, minor, moderate, severe, destroyed, unknown]
        healthServices:
          type: string
          nullable: true
          enum: [fully_functional, partially_functional, largely_disrupted, not_functioning, unknown]
        pressingNeeds:
          type: array
          items:
            type: string
            enum: [food_water, cash, healthcare, shelter, livelihoods, wash, protection, local_support, basic_services, other]
        pressingNeedsOther:
          type: string
          nullable: true
          description: Free-text detail for the `other` pressing need ("Other → specify").
...
    Report:
      type: object
      description: |-
        Canonical report document — a superset read by both clients (nested
        objects + flat aliases coexist). For public-tier callers many fields
        are stripped/coarsened; see the per-endpoint projection notes.
      properties:
        id: { type: string }
        idempotencyKey: { type: string }
        crisisId:
          type: string
          description: Empty for a pending report not yet assigned to a crisis.
        submitterId:
          type: string
          nullable: true
          description: Anonymous submitter uuid; omitted in the public projection.
        damage:
          type: string
          enum: [none, slight, moderate, severe, destroyed, minimal, partial, complete]
          description: Raw grade (3-tier OR 5-level EMS-98, depending on capture scale).
        damageTier:
          type: string
          enum: [minimal, partial, complete]
          description: The required 3-level rollup; always present.
        possiblyDamaged: { type: boolean }
        verification:
          type: string
          enum: [pending, verified, flagged]
        debris:
          type: string
          enum: ["yes", "no", "unsure"]
        infraTypes:
          type: array
          items: { type: string }
        infra:
          type: array
          items: { type: string }
          description: Dashboard alias (same array as infraTypes).
        infraName:
          type: string
          nullable: true
          description: Name/details of the infrastructure (any type), e.g. "Cumhuriyet Primary School".
        infraOtherDetail:
          type: string
          nullable: true
        crisisNature:
          type: array
          items: { type: string }
        crisis:
          type: array
          items: { type: string }
          description: Dashboard alias (same array as crisisNature).
        lat:
          type: number
          nullable: true
          description: Null for a location-unresolved (landmark-only) report — never 0,0. Coarsened to ~3 decimals (~110 m) in the public projection.
        lng:
          type: number
          nullable: true
        locationResolved:
          type: boolean
          description: True when a real GPS fix or tapped footprint produced the point.
        gpsAccuracyMeters:
          type: number
          nullable: true
        buildingId:
          type: string
          nullable: true
        buildingSource:
          type: string
          nullable: true
          description: '"footprint" when the report was pinned by tapping a real footprint polygon; null otherwise.'
        what3words:
          type: string
          nullable: true
          description: Legacy misnamed Plus Code alias — emitted with the same value as plusCode for compat.
        plusCode:
          type: string
          nullable: true
          description: Open Location Code (Plus Code) — the canonical field.
        landmark:
          type: string
          nullable: true
        place: { type: string }
        photoUrl:
          type: string
          nullable: true
          description: Relative URL of the report photo (only advertised publicly when verified).
        location:
          $ref: '#/components/schemas/ReportLocation'
        admin:
          allOf:
            - $ref: '#/components/schemas/AdminChain'
          nullable: true
        adm1Pcode:
          type: string
          nullable: true
          description: "Flat admin-1 code alias — geoBoundaries shapeID (`GB:<id>`) or seed code, NOT an OCHA P-code."
        adm2Pcode:
          type: string
          nullable: true
        adm3Pcode:
          type: string
          nullable: true
        version:
          type: integer
          description: Per-building version (server-authoritative).
        supersedesReportId:
          type: string
          nullable: true
        description:
          allOf:
            - $ref: '#/components/schemas/ReportDescription'
          nullable: true
        aiLevel:
          type: string
          nullable: true
        aiConfidence:
          type: integer
          nullable: true
        photos:
          type: array
          items:
            $ref: '#/components/schemas/PhotoRef'
        sizeBytes:
          type: integer
          format: int64
        sizeMb: { type: number }
        modular:
          allOf:
            - $ref: '#/components/schemas/ModularImpacts'
          nullable: true
        anonymization:
          $ref: '#/components/schemas/Anonymization'
        taskStatus:
          type: string
          description: "Dispatch stage: new, triaged, assigned, in_progress, resolved, closed (empty in the public projection)."
        disposition:
          type: string
          nullable: true
          enum: [resolved, cleared_nothing_found, no_action_needed, gone_on_arrival, unfounded, duplicate, referred]
        assignee:
          type: string
          nullable: true
        taskRef:
          type: string
          nullable: true
        severity:
          type: string
          description: "routine, elevated or life_safety (empty in the public projection)."
        lifeSafety: { type: boolean }
        clusters:
          type: array
          items: { type: string }
        isMine: { type: boolean }
        synced: { type: boolean }
        sync:
          description: Client sync-state blob (discriminated union), echoed as stored.
          nullable: true
        ageMin:
          type: integer
          description: Minutes since capture (computed).
        capturedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    VerificationRequest:
      type: object
      description: |-
        One of `status` (canonical) or `verification` (legacy alias) must be
        supplied; `status` wins when both are present.
      properties:
        status:
          type: string
          enum: [pending, verified, flagged]
        verification:
          type: string
          enum: [pending, verified, flagged]
          deprecated: true
          description: Legacy alias of status.
        note:
          type: string
          nullable: true
          description: Free-text analyst note, recorded in the verification audit trail.
        force:
          type: boolean
          description: Overrides the photo gate — allows status=verified on a report with no photo. Recorded in the audit trail as forced.
...